refactor: move tests of leader-followers into correct package
